Commit with tests

+++ app/src/main/java/com/openclassrooms/entrevoisins/ui/neighbour_list/NeighbourFragment.java
@@ -69,7 +69,7 @@ public class NeighbourFragment extends Fragment {
      */
     private void initList() {
         mNeighbours = mApiService.getNeighbours();
-        mRecyclerView.setAdapter(new MyNeighbourRecyclerViewAdapter(mNeighbours));
+        mRecyclerView.setAdapter(new MyNeighbourRecyclerViewAdapter(mNeighbours,0));
     }
 
     @Override
@@ -95,13 +95,18 @@ public class NeighbourFragment extends Fragment {
      *
      * @param event
      */
-    @Subscribe
-    public void onDeleteNeighbour(DeleteNeighbourEvent event) {
-        mApiService.deleteNeighbour(event.neighbour);
-        initList();
-    }
 
     @Subscribe
+    public void onDeleteNeighbour(DeleteNeighbourEvent event) {
+        if (event.fragPosition == 0) {
+            mApiService.deleteNeighbour(event.neighbour);
+            if (mApiService.getFavoritesNeighbours().contains(event.neighbour)) {
+                mApiService.deleteFavoritesNeighbours(event.neighbour);
+            }
+            initList();
+}}
+
+            @Subscribe
     public void onClickNeighbour(ClickNeighbourEvent event) {
         Intent intent = new Intent(getActivity(), NeighbourDetailActivity.class);
         intent.putExtra("Neighbour", event.neighbour);
